Removed encryption types
