+++ lib/file_transfer_agent/encrypt_util.js
@@ -9,12 +9,31 @@ const fs = require('fs');
 const Logger = require('../logger');
 const AES_BLOCK_SIZE = 128;
 const blockSize = parseInt(AES_BLOCK_SIZE / 8);  // in bytes
-const { AES_ECB, AES_GCM, AES_CBC } = require('./encryption_types');
 
 const QUERY_STAGE_MASTER_KEY = 'queryStageMasterKey';
 const BASE64 = 'base64';
 const DEFAULT_AAD = Buffer.from('');
 const AUTH_TAG_LENGTH_IN_BYTES = 16;
+
+const AES_CBC = {
+  cipherName: function (keySizeInBytes) {
+    return `aes-${keySizeInBytes * 8}-cbc`;
+  },
+  ivSize: 16
+};
+
+const AES_ECB = {
+  cipherName: function (keySizeInBytes) {
+    return `aes-${keySizeInBytes * 8}-ecb`;
+  }
+};
+
+const AES_GCM = {
+  cipherName: function (keySizeInBytes) {
+    return `aes-${keySizeInBytes * 8}-gcm`;
+  },
+  ivSize: 12
+};
 
 // Material Descriptor
 function MaterialDescriptor(smkId, queryId, keySize) {
